TS: Protein Summary - first part
